gui/requirements.txt
flask==3.0.3
flask-bcrypt==1.0.1
mysql-connector-python==9.1.0
python-dotenv==1.0.1
google-generativeai==0.3.2

